base wip: moving files around   - subtask crate   - high-level dirs:     - events     - supergraph     - subgraph     - shared   - consolidation of run_composition/runner
