+++ src/composition/supergraph/binary.rs
@@ -8,7 +8,7 @@ use camino::Utf8PathBuf;
 use tap::TapFallible;
 
 use crate::{
-    composition::{CompositionError, CompositionSuccess},
+    composition::events::{CompositionError, CompositionSuccess},
     utils::effect::{exec::ExecCommand, read_file::ReadFile},
 };
 
@@ -132,10 +132,12 @@ impl SupergraphBinary {
         let federation_version = self.get_federation_version()?;
 
         self.validate_supergraph_binary_output(supergraph_binary_output)?
-            .map(|build_output| CompositionSuccess {
-                hints: build_output.hints,
-                supergraph_sdl: build_output.supergraph_sdl,
-                federation_version,
+            .map(|build_output| {
+                CompositionSuccess::new(
+                    build_output.supergraph_sdl,
+                    build_output.hints,
+                    federation_version,
+                )
             })
             .map_err(|build_errors| CompositionError::Build {
                 source: build_errors,
@@ -177,13 +179,13 @@ mod tests {
 
     use crate::{
         composition::{
-            compose_output,
+            events::CompositionSuccess,
             supergraph::{config::FinalSupergraphConfig, version::SupergraphVersion},
         },
         utils::effect::{exec::MockExecCommand, read_file::MockReadFile},
     };
 
-    use super::{CompositionSuccess, OutputTarget, SupergraphBinary};
+    use super::{OutputTarget, SupergraphBinary};
 
     fn fed_one() -> Version {
         Version::from_str("1.0.0").unwrap()
@@ -195,6 +197,11 @@ mod tests {
 
     fn fed_two_nine() -> Version {
         Version::from_str("2.9.0").unwrap()
+    }
+
+    #[fixture]
+    fn compose_output() -> String {
+        "{\"Ok\":{\"supergraphSdl\":\"schema\\n  @link(url: \\\"https://specs.apollo.dev/link/v1.0\\\")\\n  @link(url: \\\"https://specs.apollo.dev/join/v0.3\\\", for: EXECUTION)\\n  @link(url: \\\"https://specs.apollo.dev/tag/v0.3\\\", import: [\\\"@tag\\\"])\\n  @link(url: \\\"https://specs.apollo.dev/inaccessible/v0.2\\\", import: [\\\"@inaccessible\\\"], for: SECURITY)\\n{\\n  query: Query\\n}\\n\\ndirective @inaccessible on FIELD_DEFINITION | OBJECT | INTERFACE | UNION | ARGUMENT_DEFINITION | SCALAR | ENUM | ENUM_VALUE | INPUT_OBJECT | INPUT_FIELD_DEFINITION\\n\\ndirective @join__enumValue(graph: join__Graph!) repeatable on ENUM_VALUE\\n\\ndirective @join__field(graph: join__Graph, requires: join__FieldSet, provides: join__FieldSet, type: String, external: Boolean, override: String, usedOverridden: Boolean) repeatable on FIELD_DEFINITION | INPUT_FIELD_DEFINITION\\n\\ndirective @join__graph(name: String!, url: String!) on ENUM_VALUE\\n\\ndirective @join__implements(graph: join__Graph!, interface: String!) repeatable on OBJECT | INTERFACE\\n\\ndirective @join__type(graph: join__Graph!, key: join__FieldSet, extension: Boolean! = false, resolvable: Boolean! = true, isInterfaceObject: Boolean! = false) repeatable on OBJECT | INTERFACE | UNION | ENUM | INPUT_OBJECT | SCALAR\\n\\ndirective @join__unionMember(graph: join__Graph!, member: String!) repeatable on UNION\\n\\ndirective @link(url: String, as: String, for: link__Purpose, import: [link__Import]) repeatable on SCHEMA\\n\\ndirective @tag(name: String!) repeatable on FIELD_DEFINITION | OBJECT | INTERFACE | UNION | ARGUMENT_DEFINITION | SCALAR | ENUM | ENUM_VALUE | INPUT_OBJECT | INPUT_FIELD_DEFINITION | SCHEMA\\n\\nscalar join__FieldSet\\n\\nenum join__Graph {\\n  PANDAS @join__graph(name: \\\"pandas\\\", url: \\\"http://localhost:4003\\\")\\n  PRODUCTS @join__graph(name: \\\"products\\\", url: \\\"http://localhost:4002\\\")\\n  USERS @join__graph(name: \\\"users\\\", url: \\\"http://localhost:4001\\\")\\n}\\n\\nscalar link__Import \\n\\nenum link__Purpose {\\n  \\\"\\\"\\\"\\n  `SECURITY` features provide metadata necessary to securely resolve fields.\\n  \\\"\\\"\\\"\\n  SECURITY\\n\\n  \\\"\\\"\\\"\\n  `EXECUTION` features provide metadata necessary for operation execution.\\n  \\\"\\\"\\\"\\n  EXECUTION\\n}\\n\\ntype Panda\\n @join__type(graph: PANDAS)\\n{\\n  name: ID!\\n  favoriteFood: String @tag(name: \\\"nom-nom-nom\\\")\\n}\\n\\ntype Product implements ProductItf & SkuItf\\n  @join__implements(graph: PRODUCTS, interface: \\\"ProductItf\\\")\\n  @join__implements(graph: PRODUCTS, interface: \\\"SkuItf\\\")\\n  @join__type(graph: PRODUCTS, key: \\\"id\\\")\\n  @join__type(graph: PRODUCTS, key: \\\"sku package\\\")\\n @join__type(graph: PRODUCTS, key: \\\"sku variation { id }\\\")\\n{\\n  id: ID! @tag(name: \\\"hi-from-products\\\")\\n  sku: String\\n  package: String\\n  variation: ProductVariation\\n  dimensions: ProductDimension\\n  createdBy: User\\n  hidden: String\\n}\\n\\ntype ProductDimension\\n  @join__type(graph: PRODUCTS)\\n{\\n  size: String\\n  weight: Float\\n}\\n\\ninterface ProductItf implements SkuItf\\n  @join__implements(graph: PRODUCTS, interface: \\\"SkuItf\\\")\\n  @join__type(graph: PRODUCTS)\\n{\\n  id: ID!\\n  sku: String\\n  package: String\\n  variation: ProductVariation\\n  dimensions: ProductDimension\\n  createdBy: User\\n  hidden: String @inaccessible\\n}\\n\\ntype ProductVariation\\n  @join__type(graph: PRODUCTS)\\n{\\n  id: ID!\\n}\\n\\ntype Query\\n  @join__type(graph: PANDAS)\\n  @join__type(graph: PRODUCTS)\\n  @join__type(graph: USERS)\\n{\\n  allPandas: [Panda] @join__field(graph: PANDAS)\\n  panda(name: ID!): Panda @join__field(graph: PANDAS)\\n  allProducts: [ProductItf] @join__field(graph: PRODUCTS)\\n  product(id: ID!): ProductItf @join__field(graph: PRODUCTS)\\n}\\n\\nenum ShippingClass\\n  @join__type(graph: PRODUCTS)\\n{\\n  STANDARD @join__enumValue(graph: PRODUCTS)\\n  EXPRESS @join__enumValue(graph: PRODUCTS)\\n}\\n\\ninterface SkuItf\\n  @join__type(graph: PRODUCTS)\\n{\\n  sku: String\\n}\\n\\ntype User\\n  @join__type(graph: PRODUCTS, key: \\\"email\\\")\\n  @join__type(graph: USERS, key: \\\"email\\\")\\n{\\n  email: ID! @tag(name: \\\"test-from-users\\\")\\n  totalProductsCreated: Int\\n  name: String @join__field(graph: USERS)\\n}\",\"hints\":[{\"message\":\"[UNUSED_ENUM_TYPE]: Enum type \\\"ShippingClass\\\" is defined but unused. It will be included in the supergraph with all the values appearing in any subgraph (\\\"as if\\\" it was only used as an output type).\",\"code\":\"UNUSED_ENUM_TYPE\",\"nodes\":[],\"omittedNodesCount\":0}]}}".to_string()
     }
 
     #[fixture]
@@ -211,11 +218,11 @@ mod tests {
     fn composition_output() -> CompositionSuccess {
         let res = build_result().unwrap();
 
-        CompositionSuccess {
-            hints: res.hints,
-            supergraph_sdl: res.supergraph_sdl,
-            federation_version: FederationVersion::ExactFedTwo(fed_two_eight()),
-        }
+        CompositionSuccess::new(
+            res.supergraph_sdl,
+            res.hints,
+            FederationVersion::ExactFedTwo(fed_two_eight()),
+        )
     }
 
     #[rstest]
