--- src/composition/watchers/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod subtask;
-pub mod watcher;
